chore: implemented CRUD operations in all the layouts (#2505)
* chore: basic crud operations added to the list view

* refactor: cycle details page

* refactor: module details page

* chore: added quick actions to kanban issue block

* chore: implement quick actions in calendar layout

* fix: custom menu component

* chore: separate quick action dropdowns implemented

* style: loader for calendar

* fix: build errors

+++ web/store/issue/issue_detail.store.ts
@@ -3,7 +3,7 @@ import { observable, action, makeObservable, runInAction, computed } from "mobx"
 import { IssueService, IssueReactionService } from "services/issue";
 // types
 import { RootStore } from "../root";
-import { IUser, IIssue } from "types";
+import { IIssue } from "types";
 // constants
 import { groupReactionEmojis } from "constants/issue";
 
@@ -36,17 +36,11 @@ export interface IIssueDetailStore {
   // fetch issue details
   fetchIssueDetails: (workspaceSlug: string, projectId: string, issueId: string) => Promise<IIssue>;
   // creating issue
-  createIssue: (workspaceSlug: string, projectId: string, data: Partial<IIssue>, user: IUser) => Promise<IIssue>;
+  createIssue: (workspaceSlug: string, projectId: string, data: Partial<IIssue>) => Promise<IIssue>;
   // updating issue
-  updateIssue: (
-    workspaceId: string,
-    projectId: string,
-    issueId: string,
-    data: Partial<IIssue>,
-    user: IUser | undefined
-  ) => void;
+  updateIssue: (workspaceId: string, projectId: string, issueId: string, data: Partial<IIssue>) => void;
   // deleting issue
-  deleteIssue: (workspaceSlug: string, projectId: string, issueId: string, user: IUser) => void;
+  deleteIssue: (workspaceSlug: string, projectId: string, issueId: string) => Promise<void>;
 
   fetchPeekIssueDetails: (workspaceSlug: string, projectId: string, issueId: string) => Promise<IIssue>;
 
@@ -210,12 +204,14 @@ export class IssueDetailStore implements IIssueDetailStore {
     }
   };
 
-  createIssue = async (workspaceSlug: string, projectId: string, data: Partial<IIssue>, user: IUser) => {
+  createIssue = async (workspaceSlug: string, projectId: string, data: Partial<IIssue>) => {
     try {
       runInAction(() => {
         this.loader = true;
         this.error = null;
       });
+
+      const user = this.rootStore.user.currentUser ?? undefined;
 
       const response = await this.issueService.createIssue(workspaceSlug, projectId, data, user);
 
@@ -237,13 +233,7 @@ export class IssueDetailStore implements IIssueDetailStore {
     }
   };
 
-  updateIssue = async (
-    workspaceSlug: string,
-    projectId: string,
-    issueId: string,
-    data: Partial<IIssue>,
-    user: IUser | undefined
-  ) => {
+  updateIssue = async (workspaceSlug: string, projectId: string, issueId: string, data: Partial<IIssue>) => {
     const newIssues = { ...this.issues };
     newIssues[issueId] = {
       ...newIssues[issueId],
@@ -256,6 +246,10 @@ export class IssueDetailStore implements IIssueDetailStore {
         this.error = null;
         this.issues = newIssues;
       });
+
+      const user = this.rootStore.user.currentUser;
+
+      if (!user) return;
 
       const response = await this.issueService.patchIssue(workspaceSlug, projectId, issueId, data, user);
 
@@ -282,7 +276,7 @@ export class IssueDetailStore implements IIssueDetailStore {
     }
   };
 
-  deleteIssue = async (workspaceSlug: string, projectId: string, issueId: string, user: IUser) => {
+  deleteIssue = async (workspaceSlug: string, projectId: string, issueId: string) => {
     const newIssues = { ...this.issues };
     delete newIssues[issueId];
 
@@ -293,12 +287,18 @@ export class IssueDetailStore implements IIssueDetailStore {
         this.issues = newIssues;
       });
 
-      await this.issueService.deleteIssue(workspaceSlug, projectId, issueId, user);
+      const user = this.rootStore.user.currentUser;
+
+      if (!user) return;
+
+      const response = await this.issueService.deleteIssue(workspaceSlug, projectId, issueId, user);
 
       runInAction(() => {
         this.loader = false;
         this.error = null;
       });
+
+      return response;
     } catch (error) {
       this.fetchIssueDetails(workspaceSlug, projectId, issueId);
